rb_files/requirements.txt
RPi.GPIO==0.7.1
